Move fireEventOnWrappedOrUnwrappedElement to chart utils.
