+++ ts/Accessibility/Utils/ChartUtilities.ts
@@ -17,10 +17,17 @@ import type Chart from '../../Core/Chart/Chart';
 import type { DOMElementType } from '../../Core/Renderer/DOMElementType';
 import type Point from '../../Core/Series/Point';
 import type Series from '../../Core/Series/Series';
+import type HTMLElement from '../../Core/Renderer/HTML/HTMLElement';
+import type SVGElement from '../../Core/Renderer/SVG/SVGElement';
+
 import HTMLUtilities from './HTMLUtilities.js';
 const {
     stripHTMLTagsFromString: stripHTMLTags
 } = HTMLUtilities;
+import H from '../../Core/Globals.js';
+const {
+    doc
+} = H;
 import U from '../../Core/Utilities.js';
 const {
     defined,
@@ -36,6 +43,10 @@ const {
 declare global {
     namespace Highcharts {
         interface A11yChartUtilities {
+            fireEventOnWrappedOrUnwrappedElement(
+                el: (HTMLElement|SVGElement|DOMElementType),
+                eventObject: Event
+            ): void;
             getChartTitle(chart: Chart): string;
             getAxisDescription(axis: Axis): string;
             getAxisRangeDescription(axis: Axis): string;
@@ -62,6 +73,32 @@ declare global {
 }
 
 /* eslint-disable valid-jsdoc */
+
+
+/**
+ * Fire an event on an element that is either wrapped by Highcharts,
+ * or a DOM element
+ */
+function fireEventOnWrappedOrUnwrappedElement(
+    el: (HTMLElement|SVGElement|DOMElementType),
+    eventObject: Event
+): void {
+    const type = eventObject.type;
+    const hcEvents = (el as SVGElement).hcEvents;
+
+    if (doc.createEvent && ((el as Element).dispatchEvent || (el as SVGElement).fireEvent)) {
+        if (el.dispatchEvent) {
+            el.dispatchEvent(eventObject);
+        } else {
+            (el as SVGElement).fireEvent(type, eventObject);
+        }
+    } else if (hcEvents && hcEvents[type]) {
+        fireEvent(el, type, eventObject);
+    } else if ((el as SVGElement).element) {
+        fireEventOnWrappedOrUnwrappedElement((el as SVGElement).element, eventObject);
+    }
+}
+
 
 /**
  * @return {string}
@@ -405,6 +442,7 @@ function scrollToPoint(point: Point): void {
 
 
 const ChartUtilities: Highcharts.A11yChartUtilities = {
+    fireEventOnWrappedOrUnwrappedElement,
     getChartTitle,
     getAxisDescription,
     getAxisRangeDescription,
